services/user_service/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
greenlet
python-dotenv